requirements.txt
# =========================================================
# Core Server (Daemon)
# =========================================================
fastapi==0.110.0
uvicorn[standard]==0.29.0

# =========================================================
# HTTP + Async (Client + Server communication)
# =========================================================
httpx==0.27.0
anyio==4.3.0

# =========================================================
# CLI Client
# =========================================================
typer
rich

# =========================================================
# LLM + Agents
# =========================================================
langchain
langchain-community
langgraph

# =========================================================
# Vector Memory (FAISS)
# =========================================================
faiss-cpu
sentence-transformers

# =========================================================
# Config & Environment
# =========================================================
python-dotenv

# =========================================================
# Logging
# =========================================================
loguru

# =========================================================
# Performance Optimizations
# =========================================================
orjson

